Allow defining target branch for l10n updates
This is initial implementation to allow defining a branch where the l10n
updates are pushed. This would prevent updates happening in different
OBS branches from overwriting eachother in the l10n repository master
branch.

Still needs to figure out how to define the correct branching point if
the specified branch does not exist. And to make sure that the branch
handling on the other side of the l10n machinery is ok with this.